schematic: simplify power and decoupling
The big change is switch from the lmv358a to the tlv272:

- TLV272 can operate from up to 16.5V supply, which means we can run it
  directly from the 9V battery (no 7805 or zeners required)
  - It has better typical Vos (500uV), Input Offset Current (1pA), and
    Input Bias Current (1PA)
  - It has worse max offset specs (Vos=5mV, Iio/Iib = 60pA), but they
    are still acceptable
  - Only a little more expensive (the Diodes Incorporated version on
    Digikey is $0.66 for single quantity)
  - Iio and Iib are the same, so there is no point/need for big feedback
    resistor.

Also added a lpf to the output of the amp (R3, C13) with 160hz cutoff...
the 7106 only updates at 3hz anyway, so might as well take advantage of
the low bandwidth requirements to keep noise out. (Chose 10k and 0.1uF
because I'm already using those everywhere)

Routed other bypass caps to gnd instead of -VDC for easier PCB routing.

C12 now bypasses to the negative input of U1A instead of to -VDC to make
sure high frequencies are actually being included in negative feedback.