Low battery indicator
This change adds a comparator and xor gate to implement a low-battery
indicator. When the positive rail goes below about 3V, then the output
of U5A goes high. The output of U5A is XOR'd with the backplane which
inverts the backplane when it is high, turning on the LOBAT LCD
indicator.

I'm using the spare coparator to replace the 2n7002 to keep the decimal
point on.

I got this all layed out on the PCB, but I'm increasingly unhappy with
how stringy the VDC+ trace is becoming. I think I will try laying it out
as a 4-layer board.