cheaper, smaller film caps
After reading this app note (The Integrating A/D Converter (ICL7135))
I'm convinced a big poolypropylene capacitor is not necessary. But I do
think using a film type for the autozero is a good idea.

"is of no consequence in this circuit. The dielectric absorption of
the reference and auto-zero capacitors are only important at
power-on or when the circuit is recovering from an overload.
Thus, smaller or cheaper capacitors can be used if very
accurate readings are not required for the first few seconds of
recovery."

https://www.renesas.com/us/en/document/apn/an017-integrating-ad-converter-icl7135